% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A0A1.R
\name{A0A1loc}
\alias{A0A1loc}
\title{\ifelse{html}{\out{A<sub>0</sub>}}{\eqn{A_0}},
\ifelse{html}{\out{A<sub>1</sub>}}{\eqn{A_1}} for a locus}
\usage{
A0A1loc(
  Ux,
  Uy,
  nx,
  ny,
  prob,
  pdetx,
  pdety,
  rbg,
  pfalse = 0.001,
  iminor = c(1, 1),
  plog = NULL,
  lgam = NULL,
  lnum = NULL
)
}
\arguments{
\item{Ux, Uy}{sets of unique alleles for D0 and DR samples at a given locus.
Vectors of unique indices corresponding to ordered probabilities in
\code{prob}.}

\item{nx, ny}{complexity of infection for D0 and DR samples.}

\item{prob}{a vector of population allele frequencies (on a log scale) at a
given locus.}

\item{pdetx, pdety}{detection probabilities for D0 and DR samples.}

\item{rbg}{background relatedness. Pairwise infection relatedness level in
population.}

\item{pfalse}{adjustment for possible false positive alleles allowing for
missingness when the number of unique alleles at a locus is greater than or
equal to COI. Usually a small value.}

\item{iminor}{a vector of length 2 indicating indices of minor strains in two
samples. 1 refers to a possibly recrudescent strain, 0 to unknown (any
strain can be minor).}

\item{plog}{\code{log(prob)}.}

\item{lgam, lnum}{\code{lgamma(1:x)} and \code{log(1:x)} where \code{x} is
large enough to accommodate \code{nx, ny}.}
}
\value{
\ifelse{html}{\out{log(A<sub>0</sub>),
  log(A<sub>1</sub>)}}{\eqn{log(A_0), log(A_1)}} for various combinations of
  missingness determined by \code{iminor}.
}
\description{
Calculates conditional probabilities
  \ifelse{html}{\out{P(U<sub>x, l</sub>, U<sub>y, l</sub> | R =
  0)}}{\eqn{P(U_{x, l}, U_{y, l} | R = 0)}} and
  \ifelse{html}{\out{P(U<sub>x,l</sub>, U<sub>y, l</sub> | R =
  1)}}{\eqn{P(U_{x, l}, U_{y, l} | R = 1)}} of detected unique alleles at a
  single locus for various missingness scenarios.
}
