% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{asterKM}
\alias{asterKM}
\title{Estimate therapeutic failure rate}
\usage{
asterKM(
  dsmp,
  dfppl,
  dfsmp,
  coi,
  afreq,
  rbg = 0,
  pprior = 0.5,
  update_est = TRUE,
  pfalse = 0.001,
  iminor = c(1, 1)
)
}
\arguments{
\item{dsmp}{a named list with genetic data for all the genotyped samples.
Each element represents a sample and is itself a list with elements
corresponding to loci.}

\item{dfppl}{a data frame with rows corresponding to individuals in the
study. The variables should include personal \code{id}, \code{lastday}
(last day in the study), and \code{outcome}, coded as \code{0} for
successful treatment (ACPR), \code{1} for recurrence, and \code{2} for loss
to follow-up (censoring).}

\item{dfsmp}{a data frame with rows corresponding to genotyped samples. The
variables should include personal \code{id}, \code{smp_id}, \code{day} (day
of the study the sample was obtained), and \code{pdet} (detection
probability).}

\item{coi}{a vector of COI for each sample (the length and order should
correspond to the samples in \code{dfsmp}).}

\item{afreq}{a list of population allele frequencies. Each element of the
list corresponds to a locus.}

\item{rbg}{background relatedness. Pairwise infection relatedness level in
population.}

\item{pprior}{prior probability of recrudescence.}

\item{update_est}{a logical value indicating if the failure rate estimate
should be updated once posterior probabilities are calculated.}

\item{pfalse}{adjustment for possible false positive alleles allowing for
missingness when the number of unique alleles at a locus is greater than or
equal to COI. Usually a small value.}

\item{iminor}{a vector of length 2 indicating indices of minor strains in two
samples. 1 refers to a possibly recrudescent strain, 0 to unknown (any
strain can be minor).}
}
\value{
A named list of length 2 containing:
  * Failure rate estimate;
  * Classification results for each recurrent sample along with \ifelse{html}{\out{log(A<sub>0</sub>), log(A<sub>1</sub>)}}{\eqn{log(A_0), log(A_1)}}, and posterior probabilities using a provided prior probability and empirical Bayes method.
}
\description{
Classification of recurrences and estimation of the therapeutic failure rate using Kaplan-Meier survival estimator.
}
