% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfgen}
\alias{dfgen}
\title{Allele-level data}
\format{
A data frame with 5073 rows and 3 columns. The variables \code{smp_id}, \code{locus}, and \code{allele} are character.
}
\usage{
dfgen
}
\description{
Simulated genetic data from a TES. Includes sample ID, locus, and allele.
}
\keyword{datasets}
