% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfppl}
\alias{dfppl}
\title{Person-level data}
\format{
A data frame with 97 rows and 3 columns. id column is a character vector and  lastday and outcome variables are numeric. Each row corresponds to an individual.
}
\usage{
dfppl
}
\description{
Simulated TES cohort. Includes ID, last day in the study, and outcome status.
}
\keyword{datasets}
