% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfsmp}
\alias{dfsmp}
\title{Sample-level data}
\format{
A data frame with 142 rows and 4 columns. \code{id} and \code{smp_id} variables are character and \code{day} and \code{parasite_dens} variables are numeric. Each row corresponds to a sample.
}
\usage{
dfsmp
}
\description{
Simulated list of genotyped samples from a TES. Includes individual ID, sample ID, day of the sample, and parasite density.
}
\keyword{datasets}
