% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A0A1.R
\name{PuUn}
\alias{PuUn}
\alias{Puie}
\alias{Pums}
\title{\ifelse{html}{\out{P<sub>u</sub>(U, n)}}{\eqn{P_u(U, n}}}
\usage{
PuUn(U, n, prob, plog = NULL, lgam = NULL, lnum = NULL)

Puie(U, n, prob, plog = NULL)

Pums(U, n, prob, plog = NULL, lgam = NULL, lnum = NULL)
}
\arguments{
\item{U}{a set of indices of unique alleles in a sample at a locus. The
indices should not be repeated.}

\item{n}{a number of strains to which alleles are allocated.}

\item{prob}{a vector of population allele frequencies (on a log scale) at a
given locus.}

\item{plog}{\code{log(prob)}.}

\item{lgam, lnum}{\code{lgamma(1:x)} and \code{log(1:x)} where \code{x} is
large enough to accommodate \code{nx, ny}.}
}
\value{
The value of the function.
}
\description{
\code{Puie} uses an inclusion-exclusion approach, \code{Pums} uses multisets,
and \code{PuUn} chooses a more approach based on the number of combinations.
}
