% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A0A1.R
\name{Pxy1}
\alias{Pxy1}
\title{\ifelse{html}{\out{P<sub>xy</sub><sup>(1)</sup>(U<sub>x</sub>, U<sub>y</sub>,
n<sub>x</sub>, n<sub>y</sub>)}}{\eqn{P_{xy}^{(1)}(U_x, U_y, n_x, n_y)}}}
\usage{
Pxy1(Ux, Uy, nx, ny, prob, plog = NULL, lgam = NULL, lnum = NULL)
}
\arguments{
\item{Ux, Uy}{sets of unique alleles for D0 and DR samples at a given locus.
Vectors of unique indices corresponding to ordered probabilities in
\code{prob}.}

\item{nx, ny}{complexity of infection for D0 and DR samples.}

\item{prob}{a vector of population allele frequencies (on a log scale) at a
given locus.}

\item{plog}{\code{log(prob)}.}

\item{lgam, lnum}{\code{lgamma(1:x)} and \code{log(1:x)} where \code{x} is
large enough to accommodate \code{nx, ny}.}
}
\value{
A value of the function.
}
\description{
Calculates the value of a function.
}
